+++ extensions/vscode/fidesctl-config-parser/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "fidesctl-config-parser" extension will be documented
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-
-- Initial release
+## [0.0.1] - 2021-10-06
+- Initial release, introduce basic commands 
+    - fidesctl-config-parser.fidesctlParse
+	- fidesctl-config-parser.fidesctlConfigure
+	- fidesctl-config-parser.fidesctlEnableParseOnSave
+	- fidesctl-config-parser.fidesctlDisableParseOnSave